+++ include/viua/assembler/backend/op_assemblers.h
@@ -264,6 +264,9 @@ auto assemble_op_jump(Program&,
 auto assemble_op_structremove(Program&,
                               std::vector<Token> const&,
                               Token_index const) -> void;
+auto assemble_op_structat(Program&,
+                              std::vector<Token> const&,
+                              Token_index const) -> void;
 auto assemble_op_msg(Program&, std::vector<Token> const&, Token_index const)
     -> void;
 auto assemble_op_remove(Program&, std::vector<Token> const&, Token_index const)